node interpolation uses 1/etx for weighting links olsrddatasource may use either mysqldatasource or freifunkmapdatasource for node coordinates. new parameter olsrd.nodesource = MysqlDataSource or FreifunkMapDataSource

+++ trunk/net/relet/freimap/FreiNode.java
@@ -29,7 +29,7 @@ public class FreiNode implements Comparable, Serializable {
   public String id;
   public double lon=Double.NaN, lat=Double.NaN;
   public double lonsum=0, latsum=0; //used only for real time interpolation
-  public int nc=0;
+  public double nc=0;
   public boolean unlocated=false;
   public HashMap<String, Object> attributes;
